change intervals in config from int to float64

+++ cmd/agent/flags.go
@@ -8,11 +8,11 @@ import (
 )
 
 type Config struct {
-	Addres         string `env:"ADDRESS"`
-	PollInterval   int    `env:"POLL_INTERVAL"`
-	ReportInterval int    `env:"REPORT_INTERVAL"`
-	SignKey        string `env:"KEY"`
-	RateLimit      int    `env:"RATE_LIMIT"`
+	Addres         string  `env:"ADDRESS"`
+	PollInterval   float64 `env:"POLL_INTERVAL"`
+	ReportInterval float64 `env:"REPORT_INTERVAL"`
+	SignKey        string  `env:"KEY"`
+	RateLimit      int     `env:"RATE_LIMIT"`
 }
 
 type ConfigBuilder struct {
@@ -24,12 +24,12 @@ func (b ConfigBuilder) SetAddres(address string) ConfigBuilder {
 	return b
 }
 
-func (b ConfigBuilder) SetPollInterval(interval int) ConfigBuilder {
+func (b ConfigBuilder) SetPollInterval(interval float64) ConfigBuilder {
 	b.config.PollInterval = interval
 	return b
 }
 
-func (b ConfigBuilder) SetReportInterval(interval int) ConfigBuilder {
+func (b ConfigBuilder) SetReportInterval(interval float64) ConfigBuilder {
 	b.config.ReportInterval = interval
 	return b
 }
@@ -50,11 +50,11 @@ func NewConfigFromFlags() Config {
 	var address string
 	flag.StringVar(&address, "a", ":8080", "address and port to run server")
 
-	var pollInterval int
-	flag.IntVar(&pollInterval, "p", 2, "interval between sending metric on server")
+	var pollInterval float64
+	flag.Float64Var(&pollInterval, "p", 2, "interval between sending metric on server")
 
-	var reportInterval int
-	flag.IntVar(&reportInterval, "r", 10, "interval between collecting metric from runtime")
+	var reportInterval float64
+	flag.Float64Var(&reportInterval, "r", 10, "interval between collecting metric from runtime")
 
 	var key string
 	flag.StringVar(&key, "k", "", "key for signing")
